+++ app/src/pages/Home/WatchList.tsx
@@ -49,6 +49,9 @@ export default function WatchList() {
                   <strong>Title</strong>
                 </TableCell>
                 <TableCell>
+                  <strong>Artist</strong>
+                </TableCell>
+                <TableCell>
                   <strong>Type</strong>
                 </TableCell>
                 <TableCell>
@@ -69,8 +72,13 @@ export default function WatchList() {
                   }}
                 >
                   <TableCell>
-                    <Link href={`/${row.type}/${row.id}`}>{row?.title}</Link>
+                    {row.type.includes("favorite_") ? (
+                      <>{row?.title}</>
+                    ) : (
+                      <Link href={`/${row.type}/${row.id}`}>{row?.title}</Link>
+                    )}
                   </TableCell>
+                  <TableCell>{row.artist}</TableCell>
                   <TableCell>{row.type}</TableCell>
                   <TableCell>{row.quality}</TableCell>
                   <TableCell>