+++ app/src/provider/ProcessingProvider.tsx
@@ -10,9 +10,12 @@ import { EventSourceController } from "event-source-plus";
 import {
   AlbumType,
   ArtistType,
+  ContentType,
+  FavoritesType,
   MixType,
   PlaylistType,
   ProcessingItemType,
+  SyncItemType,
   TrackType,
   VideoType,
 } from "../types";
@@ -31,8 +34,9 @@ type ProcessingContextType = {
         | ArtistType
         | PlaylistType
         | MixType
-        | VideoType,
-      type: "album" | "track" | "artist" | "playlist" | "video" | "mix",
+        | VideoType
+        | FavoritesType,
+      type: ContentType,
     ) => void;
     removeItem: (id: string) => void;
     retryItem: (item: ProcessingItemType) => void;
@@ -60,8 +64,9 @@ export function ProcessingProvider({ children }: { children: ReactNode }) {
       | ArtistType
       | PlaylistType
       | MixType
-      | VideoType,
-    type: "album" | "track" | "artist" | "playlist" | "video" | "mix",
+      | VideoType
+      | FavoritesType,
+    type: ContentType,
   ) => {
     if (!quality) return;
 
@@ -80,7 +85,7 @@ export function ProcessingProvider({ children }: { children: ReactNode }) {
       id: id,
       artist:
         type === "artist"
-          ? (item as ArtistType)?.name
+          ? (item as ArtistType)?.name || (item as SyncItemType)?.artist || ""
           : (item as TrackType | AlbumType).artists?.[0].name,
       title:
         type === "artist"